requirements.txt
Complete output from command git config --get-regexp remote\..*\.url:

----------------------------------------
alabaster==0.7.12
awscli==1.16.38
Babel==2.6.0
backcall==0.1.0
botocore==1.12.28
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.3.9
coverage==4.5.1
cycler==0.10.0
decorator==4.3.0
docutils==0.14
flake8==3.5.0
idna==2.7
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
missingno==0.4.1
numpy==1.15.2
packaging==18.0
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.6
ptyprocess==0.6.0
pyasn1==0.4.4
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pyparsing==2.2.2
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
requests==2.20.0
rsa==3.4.2
s3transfer==0.1.13
scipy==1.1.0
seaborn==0.9.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
## !! Could not determine repository location
src==0.1.0
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24
wcwidth==0.1.7
